fix: swap input access view for ios
